In Row and Column Enumerators, an addition for if the enumeration does not start from [x,0] and [0,y] respectively.
The Sparse Matrix keeps track of its bounds itself, the benefit is now there is no looping every time the bounds are asked for. The selection (with its mass renaming Top->top) is now a FungeSpaceArea.

Finally, there is a start to be able to move the view screen around! Tab + An arrow key will move it around. A bug that causes an infinite loop is for fs_view_screen to have its top left corner moved past the bounds of the matrix.