This very cool update works on selection and drawing. The way selection is updated and moved is now much more simple, and its "active" variable is removed leading to a lot less "bad state" bugs. The EditIP is no longer drawn when selecting. When selecting over the right or bottom edge the view pane will scroll allowing to select any number of lines! Select works in all quadrants.
You can now paste in any Unicode character you want, which is handy as alt code input doesn't seem to work right now.

A running IP now draws where the next command that will be executed is, regardless of if it is after a blank ;section; or over an edge to wrap around. Fixed a bug where "1;    ;    2" would not skip over all the space between 1 and 2.